feat: move terraform_drift_detection config to top level script
